fix for "unknown author" when a post is required to see replies
"Unknown Author" typically shows up when you fetch a discussion
before the background jobs have processed it. In this case, we
were fetching the discussion immediately after adding the
student's reply, so the background jobs didn't have a chance to
get to the new reply yet.

Normally this isn't a problem because the reply is added to the
discussion without refetching the discussion. In this case, we
need to remember the new reply, which has all the good
information in it, so we can replace bad fetched reply with the
good data.

fixes #CNVS-2406

test plan:
 - As a teacher, create a discussion topic with "Users must
   post before seeing replies" checked
 - As a teacher, create reply to the discussion
 - View the discussion as a student (shouldn't see replies)
 - Reply to the discussion as a student
 - Replies should load and you should see student's reply with
   the student's name

Change-Id: Id1189eabc0e8e1ff90c3764f2d0d92b425b50672
Reviewed-on: https://gerrit.instructure.com/16413
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
QA-Review: Marc LeGendre <marc@instructure.com>