user-in-course cross-sharding bug
fixes #CNVS-1831

context_controller was calling
@context.users.find(params[:id]).  When that
id was a long id (like one for a copied user
from another shard) it would not find the user
in that collection because, although the user 
would be loaded in the "users" collection, it's id
would be the normal local id (not the global one
being passed in params[:id]).  using
the enrollment which had already been loaded
to pull the user from solves the problem.

Also, had to read through user merging in the
process, and found it too big to really belong
in the user model, so refactored it out into 
it's own class in the lib directory.

TEST PLAN:
1) start with 2 users enrolled in 2 different
courses on 2 different shards.

2) merge the users together.

3) as an admin, go to the course on the
shard that the user was deleted from (not
the shard where the user remained) and
try to click through the course's user roster
to that user.

4) You should not be stopped by a message 
saying the user does not exist.


Change-Id: I902c37b4a2c7d8ea010cee046700a209cb99536a
Reviewed-on: https://gerrit.instructure.com/16869
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ethan Vizitei <ethan@12spokes.com>
QA-Review: Clare Hetherington <clare@instructure.com>