gradebook history api
fixes #CNVS-2802

This provides a reimplementation of 
the functionality in 'lib/submission_list.rb'.
That lib file has been used to produce
a grade book history page, but it loads
the entire history of the course at once,
which is untenable.  This take all the same data,
and provides it through a paginated
API endpoint.  Using this API, a rich front
end will be possible to create in order to replace
the current large grade book history page, 
one that makes API calls to reveal data further d
own the tree as it is requested.  This is an 
unusual API endpoint in that it does not present 
data as it is in the database, there is a series 
of transformations that the submission data is
put through to arrive at a versioned history
where each node contains within itself some
contextual knowledge of the flow of the
submission progress (each version knows
what the grade of the version before it was, 
and what the grade is today, for example).

TEST PLAN:
This is a new API endpoint and does not
yet get used by any front end code yet.
It can be played with by performing some
valid API queries against the following paths:

/courses/:course_id/gradebook_history/days
/courses/:course_id/gradebook_history/:date 
^ date is like "2013-01-31" ^
/courses/:course_id/gradebook_history/:date/graders/:grader_id/assignments/:assignment_id/submissions
(yes, that last one is huge, but it does follow
the nested structure described by the original
grade book history page).

The user that is selected for testing needs to 
have the "manage_grades"
permission on the referenced course.  API
documentation is available on the controller class
app/controllers/gradebook_history_api_controller.rb

Change-Id: I18e0b4b967d6c20ad47b86e98adbc15b87276098
Reviewed-on: https://gerrit.instructure.com/17366
QA-Review: Clare Hetherington <clare@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>