use a custom Marshal format for Time objects pre-1900 in ruby 1.9
Ruby 1.9 changed the Marshal format for Times and now raises an
exception for Time objects with a year < 1900. This extends _dump and
_load to handle these older Times.

test plan: An easy test in the UI is to (with caching enabled) create an
enrollment term whose start or end date is before 1900, and restrict
student enrollments to that date. Then visit the dashboard and various
course pages as a student in the course, and verify you don't see page
errors.

Change-Id: I97925925e04bb3663efe9928810845c6c919adac
Reviewed-on: https://gerrit.instructure.com/17811
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>