add resource link id to assignments api
FIXES #CNVS-3625

grabbed the opaque_identifier for the tag
because that's what was being used in 
'ApplicationController#content_tag_redirect'.

TEST PLAN:

1) use the API to access an assignment
that has an external_tool_tag

2) you should get "external_tool_tag_attributes"
as an attribute in the assignment json. It should
have a field called "resource_link_id" in it
with a long id value.


Change-Id: I983dc12c97cc65eaea97638364e3a68f0f279a9f
Reviewed-on: https://gerrit.instructure.com/17867
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Adam Phillipps <adam@instructure.com>