don't even attempt validating a blank password
fixes CNVS-4447

test plan: the javascript on the normal login page blocks blank
passwords, so the easiest way to test this currently is to use a mobile
browser or change your user-agent to be a mobile browser so that you get
the mobile login screen, which currently allows blank passwords through
(though that'll be fixed at some point too). though canvas won't allow a
blank password, so you'll also need an LDAP server that is configured to
have your password as the blank string.

Change-Id: Ie619638c4c6362461f2bebc5417ef6397d0d9ad1
Reviewed-on: https://gerrit.instructure.com/18387
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Paul Hinze <paulh@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>