fix attachment.open with a block
This broke with the conversion to aws-sdk

test plan: do a course import that has text or html files in the package
(i discovered this while doing an Angel import). creating the
attachments in the course should not fail. (before this commit it was
failing in infer_encoding)

Change-Id: Ia492f5e42436e2db2b6c66852264af8b1731499c
Reviewed-on: https://gerrit.instructure.com/18652
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>