force health check to honor the TMPDIR env var if set
Dir.tmpdir (and thus Tempfile) has some internal checks to make sure
that the directory specified in ENV['TMPDIR'] exists and is writable,
otherwise it falls back to the system default (usually /tmp). But some
other parts of the stack such as Passenger/Rack still try to use TMPDIR
even if it's not writable, which can happen if it's a bad drive, so we
want to force the health check to always try the directory anyway.

test plan: /health_check should return "ok" still if TMPDIR is set and
writable. It should fail if TMPDIR is not writable, even if /tmp is
writable.

Change-Id: I6d4efed9327dc6fb6395940cd90bd5e93d7c4069
Reviewed-on: https://gerrit.instructure.com/18691
Reviewed-by: Cody Cutrer <cody@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>