visual indicators for answered quiz questions
FIXES CNVS-4942

The problem statement for this ticket was that
answered questions could only be seen as
answered in the right column because of the
color change, which is insufficient for
accessibility.

This patch uses icons to differentiate between
answered and unanswered questions.

TEST PLAN:

1) login as a student to a course where you
have a quiz to take

2) while taking the quiz, look at the right
column

3) the icon for questions you've already answered
should be a checkmark

Change-Id: I5ec289acda00aa9e14ef49e387c8756883424c0a
Reviewed-on: https://gerrit.instructure.com/19456
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Dana Morley <dana@instructure.com>