app center external apps endpoint
provides an unpublished, backbone accessible api endpoint for the app center to consume.
In this commit the apps api just acts as a proxy to the external apps website based on plugin
settings, but in the future this endpoint will be able to match up data to tools installed in the
current context, cache data from the external website, return data to the external website,
and so forth.

Test Plan:
  - Before installing the plugin:
    * Visit /api/v1/accounts/:account_id/app_center/apps
    * You should get an empty json array
  - Enable the app center plugin
    * Visit /api/v1/accounts/:account_id/app_center/apps
    * You should get a json formatted array of app data

Change-Id: I48072053950e67aa583a3c79e2a582d30e7cd409
Reviewed-on: https://gerrit.instructure.com/20161
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>
QA-Review: Bracken Mosbacker <bracken@instructure.com>