bind selenium thin server to 0.0.0.0 by default
the selenium thin server was using getaddress/gethostname to choose
an address to bind to. this was intermittently prevening the browser
from connecting to the server because the bind ip was selected
differently than the ip given to the browser.

fixes #CNVS-5761

Change-Id: I759f667d3eb677014faf5a539434849a8c7e58d2
Reviewed-on: https://gerrit.instructure.com/20422
Reviewed-by: Bryan Madsen <bryan@instructure.com>
Product-Review: Bryan Madsen <bryan@instructure.com>
QA-Review: Bryan Madsen <bryan@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>