Draft State for Modules
Fixes CNVS-6294

This adds the same published button for modules
as we had for the quiz display page, and then
abstracts the two into using the PublishButton
backbone view.  A publishable model has been
created to represent anything that is a publishable
entity (because adding the functionality to existing
models was proving to be longer than expected).
Perhaps in the future taking the publishable
model and making it a mixin for all actual
models that need that functionality would be best.

TEST PLAN:

1) Login as an instructor
2) Regression test the quiz publish button
3) navigate to the modules page for a course
4) you should see publish/published buttons.
5) clicking on any publish button for a module
should toggle it's state.

Change-Id: I2639354c834af3aeabd3d0e9e84e2c70504961cf
Reviewed-on: https://gerrit.instructure.com/21499
Reviewed-by: Stanley Stuart <stanley@instructure.com>
Product-Review: Ethan Vizitei <ethan@12spokes.com>
QA-Review: Ethan Vizitei <ethan@12spokes.com>
Tested-by: Jenkins <jenkins@instructure.com>