fix sharding issue in per-user authentication logs
fixes CNVS-7452

User#active_pseudonyms will query on User#shard regardless of
Shard.current. But the value yielded to a partition_by_shard block is a
set of local_ids. So e.g. if visible_accounts included an account from
shard2 but @user is on shard1, we'd look for a pseudonym on shard1
belonging to both the user and an account on shard1 with the same local
id, rather than looking for a pseudonym on shard2 belonging to both the
user and the intended account.

the majority of the time, this bug would only cause the pseudonym to not
be found (in cooperation with another bug, this caused CNVS-7452). but
this could also provide unintended access, which is more severe. for
example:

 * user1 and account1 are on shard1
 * account2 is on shard2 but has the same local id as account1
 * user1 has pseudonyms in both account1 and account2
 * user2 has admin privileges in account2 but not account1
 * user2 requests user1's authentication log
 * user2 should get the events for the account2 pseudonym
 * user2 gets the events for the account1 pseudonym instead

test-plan:
 - setup user1, user2, account1, account2, and pseudonyms as in the
   example above
 - generate some authentication events for both pseudonyms
 - user2 should get the events for the account2 pseudonym
 - user2 should not get any events for the account1 pseudonym

Change-Id: Icc639d7a4b8217b77bd859629f983fbc1d30ccc0
Reviewed-on: https://gerrit.instructure.com/23204
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>