[WebAIM] quiz answer indicators not read by SR
Quiz answers should be read by the screen reader when their
corresponding element receives focus.

closes CNVS-6634

Test plan:

  - As a Teacher, create a quiz with questions of (preferably) all types
  - Make sure the quiz has one "Fill in the blanks" question as it
    represents an edge case
  - Publish the quiz
  - As a Student, take the quiz, and get some answers right!
  - Using JAWS or a SR, advance to the radio buttons or text fields
  - Confirm that the SR speaks "Correct", "You Answered" appropriately

Notes:

When viewing the quiz attempt, the display of answers of some question types
are not consistent (namely, FITB and Numerical questions) in that the
correct answer is set inside a text field (<input />) while the other
answers (or the student-supplied one) are set in nodes of different
types (<div />). This causes these other answers _not_ to receive focus
by the screen reader.

However, this is not related to this issue (only that it affects it.)

Change-Id: I12c38d22768fdc759b5a3719e91c03d78b8fca90
Reviewed-on: https://gerrit.instructure.com/24300
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>