discussions: allow subscribing after reply to topic requiring reply
fixes CNVS-8019

The ticket says that IE9 used to refresh the page when adding a reply to
a topic with "Users must post before seeing replies" turned on and that
it was useful because it displayed the subscription button. Instead of
restoring that behavior, this change just displays the subscription
button on the client side after adding a reply, so all browsers can see
it.

Test plan:
  * open a topic with "Users must post before seeing replies" enabled
      as a student who has not replied
  * add a reply
  * verify that, in addition to seeing the other replies, you now
      immediately see the subscription button

Change-Id: Ieb91938d2f320f996c1d2d48ca5eb26675c4e465
Reviewed-on: https://gerrit.instructure.com/24353
QA-Review: Cam Theriault <cam@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Willesen <jonw@instructure.com>
Product-Review: Braden Anderson <banderson@instructure.com>